feat(document): implémente la préviz des document PDFs dans un nouvel onglet
