+++ package.json
@@ -41,8 +41,7 @@
     "connect-history-api-fallback": "^1.6.0",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "http-proxy-middleware": "^2.0.1",
-    "pdfjs-dist": "^2.10.377"
+    "http-proxy-middleware": "^2.0.1"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.16.5",
