+++ src/styles/system/ui.css
@@ -108,7 +108,6 @@
 
 .border-b-s:not(:last-child) {
   border-bottom: var(--unit-px) solid var(--color-alt);
-
   /* margin-bottom: calc(-1 * var(--unit-px)); */
 }
 
@@ -146,6 +145,7 @@
   text-align: center;
   border: var(--border-width) solid var(--border-color);
   border-radius: var(--unit-s);
+
   position: absolute;
   left: 150%;
   z-index: 1;
@@ -196,20 +196,4 @@
 .fixed-width {
   width: calc(var(--col-width) * 1 / 4);
   display: inline-block;
-}
-
-.pdf-visualizer-container {
-  position: relative;
-  display: block;
-  width: 100%;
-  height: 100%;
-  z-index: -1;
-}
-
-.pdf-visualizer {
-  width: 100%;
-}
-
-.pdf-visualizer-pagination {
-  margin-top: -10px;
 }
