--- src/typings/pdfjs.d.ts
@@ -1,1 +1,0 @@
-declare module 'pdfjs-dist/build/pdf.worker.entry'
